+++ app/voyage/Container.tsx
@@ -1,6 +1,9 @@
 'use client'
 
-/* Server rendered client components here */
+/**
+ * This component is a client component to hold state but does
+ * not require the maplibre instance
+ **/
 
 import { useMemo, useState } from 'react'
 
@@ -25,33 +28,6 @@ import useTransportStopData from './transport/useTransportStopData'
 const Map = dynamic(() => import('./Map'), {
 	ssr: false,
 })
-
-export async function generateMetadata(
-	{ params, searchParams }: Props,
-	parent: ResolvingMetadata
-): Promise<Metadata> {
-	console.log('will METADATA')
-	const allez = searchParams.allez.split('|')
-
-	if (!allez.length) return null
-	const vers = allez[allez.length - 1]
-	const step = stepOsmRequest(vers)
-	if (!step) return null
-
-	const title = step.name || step.osmFeature?.name,
-		description = 'blabla',
-		image = `https://cavedupalais.shop/cdn/shop/files/grande-chartreuse-verte-55-35cl_460x@2x.jpg?v=1683104268`
-
-	const metadata = {
-		title: title,
-		description,
-		openGraph: {
-			images: [image],
-		},
-	}
-	console.log('METADATA', metadata)
-	return metadata
-}
 
 export default function Container({ searchParams }) {
 	const setSearchParams = useSetSearchParams()
